Fix code based on code review.
