+++ Sources/Components/TimeLineView/SimpleTimeLineIndicatorProvider.swift
@@ -31,57 +31,3 @@ public struct SimpleTimeLineIndicatorProvider: TimeLineIndicatorProvider {
         )
     }
 }
-
-final class SimpleTimeLineIndicatorView: UIView {
-    static var indicatorDotSize: CGFloat = 8
-
-    private var dotLayer: CAShapeLayer?
-
-    var dotOffset: CGFloat = 0
-
-    override func layoutSubviews() {
-        dotLayer?.removeFromSuperlayer()
-
-        let indicatorDotY = dotOffset + SimpleTimeLineIndicatorView.indicatorDotSize / 2
-        dotLayer = layer.addCircle(
-            at: CGPoint(
-                x: bounds.midX,
-                y: indicatorDotY
-            ),
-            radius: DottedTimeLineIndicatorView.indicatorDotSize / 2,
-            color: .timelineIndicator
-        )
-    }
-
-    override var intrinsicContentSize: CGSize {
-        return CGSize(
-            width: SimpleTimeLineIndicatorView.indicatorDotSize,
-            height: UIView.noIntrinsicMetric
-        )
-    }
-
-}
-
-extension CALayer {
-    public func addCircle(at position: CGPoint, radius: CGFloat, color: UIColor) -> CAShapeLayer {
-        let circlePath = UIBezierPath(
-            arcCenter: position,
-            radius: radius,
-            startAngle: 0,
-            endAngle: .pi * 2,
-            clockwise: true
-        )
-
-        let shapeLayer = CAShapeLayer()
-        shapeLayer.path = circlePath.cgPath
-        shapeLayer.fillColor = color.cgColor
-        addSublayer(shapeLayer)
-        return shapeLayer
-    }
-}
-
-// MARK: - Private extensions
-
-private extension UIColor {
-    static var timelineIndicator = dynamicColor(defaultColor: .sardine, darkModeColor: .darkSardine)
-}
